LPD-88102 Move ExportPreview mock to test fixtures folder
